requirements.txt
requests==2.32.4
streamlit==1.48.0
fastapi==0.116.1
pydantic==2.11.7
pydantic-settings==2.10.1
uvicorn==0.35.0
llama-index==0.13.1
llama-index-llms-ollama==0.7.0
llama-index-llms-groq==0.4.0
llama-index-llms-openai==0.5.2
python-dotenv==1.1.1
tinydb==4.8.2